+++ Frontend/CHANGELOG.md
@@ -1,8 +1,16 @@
 # ChangeLog 
->Please update this to let other developers whats different, 
-we uses Semantic Version
+>Please update this to let other developers know whats going on with all this busy-ness, 
+ans we uses Semantic Version.
 
 ##unreleased
+
+##version 0.2.0 - 3/20/2017
+
+[code] Created a delete method for `companies.service`
+[code] finished the `companies.service` CRUD
+[code] added a modal library and updated the `contact.services`
+
+##version 0.1.1 - 3/20/2017
 
 [refactor] removed top navbar and option button from `home.component`
 [code] added POST and GET to `company.service`